Added Finish() to SimpleProcessor interface, and converted some tests.
The Finish() function gets called after when the logline channel gets
closed, but before closing the downstream channel. This is especially
useful for end-of-line processors that need to do something with the
results.

Existing processors have been updated, and the builder tests have been
converted to use SimpleProcessors which cleans things up a bit.