Added preprocessors and cusome DataCollector to the yaml conf.
Preprocessors are implemented as their own separate pipeline, with the
input forked. They may emit data at any rate, and their data is merged
back with the main input stream via Timeweavers.

The custom DataCollector can be used to forward LastHop pipeline data to
a custom object, which can be especially useful when you want to reuse a
processor but change the persistence model.