Simpilfied FilterConf.
Instead of having SimpleFilter and ComplexFilter, we now have one
FilterSpec with a Type field that can be simple, custom, or regex. The
Filter field is a single string and for "simple" this is a "&&"
delimited string. For "custom", this is the name of the existing,
registered function. For "regex", it's the regex string.